requirements.txt
annotated-types==0.7.0
apispec==6.8.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
black==25.1.0
blinker==1.9.0
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
coverage==7.10.3
Deprecated==1.2.18
distlib==0.4.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.18.0
flake8==7.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
flask-smorest==0.46.1
flask-talisman==1.1.0
gunicorn==23.0.0
identify==2.6.13
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mongomock==4.3.0
mypy_extensions==1.1.0
nodeenv==1.9.1
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.0
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rich==13.9.4
sentinels==1.0.0
structlog==25.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
virtualenv==20.33.1
webargs==8.7.0
Werkzeug==3.1.3
wrapt==1.17.2
